+++ wemake_python_styleguide/visitors/ast/complexity/nested.py
@@ -56,7 +56,7 @@ class NestedComplexityVisitor(BaseNodeVisitor):
     def _check_nested_lambdas(self, node: ast.Lambda) -> None:
         parent = getattr(node, 'parent', None)
         if isinstance(parent, ast.Lambda):
-            self.add_violation(NestedFunctionViolation(node))
+            self.add_violation(NestedFunctionViolation(node, text='lambda'))
 
     def visit_ClassDef(self, node: ast.ClassDef) -> None:
         """