+++ wemake_python_styleguide/visitors/presets/complexity.py
@@ -12,13 +12,19 @@ from wemake_python_styleguide.visitors.ast.complexity import (
 #: Used to store all complexity related visitors to be later passed to checker:
 COMPLEXITY_PRESET = (
     function.FunctionComplexityVisitor,
+
     jones.JonesComplexityVisitor,
+
     nested.NestedComplexityVisitor,
+
     offset.OffsetVisitor,
+
     counts.ImportMembersVisitor,
     counts.ModuleMembersVisitor,
     counts.MethodMembersVisitor,
     counts.ConditionsVisitor,
     counts.ElifVisitor,
+    counts.TryExceptVisitor,
+
     classes.ClassComplexityVisitor,
 )