+++ wemake_python_styleguide/visitors/ast/complexity/nested.py
@@ -37,14 +37,14 @@ class NestedComplexityVisitor(BaseNodeVisitor):
     )
 
     def _check_nested_function(self, node: AnyFunctionDef) -> None:
-        parent = getattr(node, 'parent', None)
+        parent = getattr(node, 'wps_parent', None)
         is_inside_function = isinstance(parent, self._function_nodes)
 
         if is_inside_function and node.name not in NESTED_FUNCTIONS_WHITELIST:
             self.add_violation(NestedFunctionViolation(node, text=node.name))
 
     def _check_nested_classes(self, node: ast.ClassDef) -> None:
-        parent = getattr(node, 'parent', None)
+        parent = getattr(node, 'wps_parent', None)
         is_inside_class = isinstance(parent, ast.ClassDef)
         is_inside_function = isinstance(parent, self._function_nodes)
 
@@ -54,7 +54,7 @@ class NestedComplexityVisitor(BaseNodeVisitor):
             self.add_violation(NestedClassViolation(node, text=node.name))
 
     def _check_nested_lambdas(self, node: ast.Lambda) -> None:
-        parent = getattr(node, 'parent', None)
+        parent = getattr(node, 'wps_parent', None)
         if isinstance(parent, ast.Lambda):
             self.add_violation(NestedFunctionViolation(node))