+++ wemake_python_styleguide/visitors/ast/imports.py
@@ -32,7 +32,7 @@ class _ImportsValidator(object):
         self._error_callback = error_callback
 
     def check_nested_import(self, node: AnyImport) -> None:
-        parent = getattr(node, 'parent', None)
+        parent = getattr(node, 'wps_parent', None)
         if parent is not None and not isinstance(parent, ast.Module):
             self._error_callback(NestedImportViolation(node))