+++ wemake_python_styleguide/visitors/ast/keywords.py
@@ -103,11 +103,11 @@ class WrongComprehensionVisitor(BaseNodeVisitor):
         if len(node.ifs) > self._max_ifs:
             # We are trying to fix line number in the report,
             # since `comprehension` does not have this property.
-            parent = getattr(node, 'parent', node)
+            parent = getattr(node, 'wps_parent', node)
             self.add_violation(MultipleIfsInComprehensionViolation(parent))
 
     def _check_fors(self, node: ast.comprehension) -> None:
-        parent = getattr(node, 'parent', node)
+        parent = getattr(node, 'wps_parent', node)
         self._fors[parent] = len(parent.generators)
 
     def _post_visit(self) -> None: