+++ CHANGELOG.md
@@ -4,6 +4,13 @@ We follow Semantic Versions since the `0.1.0` release.
 We used to have incremental versioning before `0.1.0`.
 
 
+## WIP
+
+### Misc
+
+- Adds `safety` and other dependency checks to CI process
+
+
 ## Version 0.6.0
 
 ### Features